Refactor naming, imports, sorting. Part 5
