+++ src/components/abstract/component-smart.js
@@ -1,6 +1,6 @@
-import {AbstractComponent} from "./component.js";
+import AbstractComponent from "./component.js";
 
-class AbstractSmartComponent extends AbstractComponent {
+export default class AbstractSmartComponent extends AbstractComponent {
   recoveryListeners() {
     throw new Error(`Abstract method not implemented: recoveryListeners`);
   }
@@ -17,6 +17,3 @@ class AbstractSmartComponent extends AbstractComponent {
     this.recoveryListeners();
   }
 }
-
-
-export {AbstractSmartComponent};
